first, untested implementation
